Re-introduce Archive as the root object
I'm not sure how to describe what's going on, but the Archive is
responsible for looking like an ordinary directory, but operates in such
a way that every time we work with an Archive, we're working with a
clean slate. When we want to turn the Archive into an IO stream, the
Archive merges all of its clean slates together to form the final
picture.